Write a new packaging script for Dafny
This is intended to superseed the PrepareDafnyZip.bat script. I tried doing this
as an msbuild config, then as a makefile, but the first one was too cumbersome
and the second one required too many dependencies to work properly (and parsing
JSON from a makefile is not fun).

The new script calls the Github API to retrieve the description of the latest z3
release, downloads the zip corresponding to each supported platfom, and
packages one copy of Dafny for each copy of z3 that the latest release
contains.

The next step will be to make Dafny load its binary from the z3 folder.