Make the consequence axiom trigger off of any fuel, not just StartFuel_F.
dafny4/FlyingRobots.dfy was failing because the consequence axiom for Valid was triggered off of StartFuel__module.Bot.Valid, whereas the ensures for the constructor mentions StartFuelAssert__module.Bot.Valid.  Without a reveal (or a well crafted assert), we don't know that those are related to each other.

As a result of this change, dafny4/ACL2-extractor.dfy and dafny2/SnapshotableTrees.dfy now pass.