Added "allocated" expressions, which say that the argument expression, which can be of any type, is available in the state.  The "allocated" expression is typically used with "old", as in "old(allocated(E))" (note that "E" also sits inside the "old" in this example).
Changed "fresh" so that no longer is allowed to take a datatype argument (which was kinda bogus
anyway).  Where "!fresh(D)" was used before, try "old(allocated(D))" instead.

Changed semantics of "fresh", so that its value on a (set/seq) collection is simply the fold
of "fresh" on all of the elements.  In particular, this means that if the collection contains
"null", then "fresh" will return "false".  This is likely to affect some classes specified with
almost-the-standard dynamic-frames idiom -- make sure "Valid()" includes a conjunct "null !in Repr".

Fixed bug in the verification translation of local variables without initializing assignments.

Improved triggers in some "fresh(S)" expressions, where "S" is a set of references.