Don't empty methodbody of lemma from include file while resolving
Methodbody of lemma from include files is emptied in pre-resolve rewrite
phase in order to speed up resolver and translator. This causes a
problem when a module refines another module that is defined in the
include file since the rewriter would empty the methodbody for lemma
defined in include file. Instead of using rewrite to empty the
methodbody, simply just stop resolve and translator methodbody if a
lemma is from included file. Closes #46.