CompileSignature for ModuleExportDecl was set wrong.
After cloning module definition before second pass resolve, we set the
ModuleExportDecl's compileSignature to the cloned module's signature,
which is wrong since we could be exporting only part of the module.
Instead, we should update the ModuleExportDecl's signature to point to
the cloned modules. Also, when importing from more than one exports, we
always set the compileSignature to the last export's compileSignature,
which is wrong. We need to merge the compileSignatures instead. Closes
#50.