Fix incorrect translation of forallstmt
1. For forall proof, we used to use initEtran to translate boundvars and
range and etran to translate the Ens. To switch between initEtran and
etran when translating the corresponding forallexpression of the
forallproof, we made initEtran as the Old of etran. However, with the
commit# bdf8eb3, we only need etran therefore making the code that
switch between etran and initEtran not needed for forallproof.
2. For forallstmtcall, there are code for handling translating the stmt
when there is no corresponding forallexpression. In stead of  using
TrExpr for forallstmtcall that has forallexpression, we use the same
code as the first case, and then add the triggers from forallexpression.
3. We quantify over heap for calls that don't use heap so quantifier can
preserve across method calls since we always include the heap as an
argument even in non-heap reading functions (see
test\dafny4\bug144.dfy). However, we don't need to quantify over heap
for calls that uses heap.
4. Fix forallexpression rewrite for forallstmtcall so that triggers can
be generated.
Closes #67.