Rewrite quantifiers to avoid matching loop.
When matching loops is detected in a quantifier during tigger selection,
the quantifier will be split along conjunctions and each resulting
quantifier will be rewritten and then trigger select independently.
Also, add  "matchinglooprewrite" attribute so that we can disable
rewriting the quantifier when the automatic induction in a test is not
powerful enough to instantiate induction hypothesis. Closes #69.