Changed syntax of arrow types. There are now built-in arrow types with a built-in syntax for its arrow:
A ~> B
This is the most general arrow. It allows the function to have read effects (that is, have a reads clause) and allows it to be partial (that is, have a requires clause). This arrow was previously supported, but with the syntax A -> B.

A --> B
This is the partial arrow. It does not allow read effects, but it allows the function to have a precondition.  The type is a built-in subset type of the general arrow, that is, it is declared as follows (where I'm using its syntax and imagining that A and B are type parameters).
type A -> B = f: A ~> B | forall a :: f.reads(a) == {}

A -> B
This is the total arrow. It does not allow either read effects or preconditions. Note that the previous syntax -> now means a total arrow, not the general arrow. The total arrow is a built-in subset type of the partial arrow:
type A -> B = f: A --> B | forall a :: f.requires(a)

In the illustrations above, the arrow type show has arity 1, but all three arrows are supported for any arities.