Updates to make Mono build work with VS2017 features + contracts.
Now that we're using VS2017 features, it seems reasonable to require a
modern version of Mono (the version in Ubuntu repos is several years out
of date). This includes switching from xbuild to msbuild, which can
handle new VS2017 project files.

Also, the build now uses Code Contracts, but those don't work well with
Mono, so only try to use them if they are actually installed.