Implement simplification pass before translation to Boogie
This adds a simplification pass based on lemmas with a :simp attribute before
the program is translated to Boogie. Any call to a function with the {:simp}
attribute will cause arguments to the function to be rewritten using lemmas
with the :simp attribute from left to right. Note that this can introduce loops
depending on the simplification lemmas.