Support for local simplification rules in simplifier
This implements preliminary support for keep track of
simplification results assigned to variables earlier in the
method; concretely, for an assignment var x := simp(e);
we add a rewrite rule x ~> e' where e' is the result of simplifying
e.

However, this is currently not sound since we need to invalidate
these rules whenever any variable in e' gets assigned to
or shadowed or otherwise modified.