Go: Modules, function calls
Calls.dfy now passes the lit test extended for Go, in addition to
Hello.dfy.

Modules must be in separate files in Go, so this required setting up a
system for multi-file output.  In particular, ToString() on
TargetWriters is now sketchy, though it should (!) be okay for now on
small fragments that won't introduce modules.  (I'd put an Obsolete
attribute on TargetWriter.ToString(), but C# can't handle marking an
override Obsolete.)

One other change to the base Compiler class was in generation of static
fields and methods.  There are no static members in Go, so *all* classes
(not just traits) need companion objects.  Also, methods and fields aren't
defined in the same body, so we need to have separate TargetWriters for
static fields vs. static methods, which we didn't need before Go.

Finally, the handling of out parameters needs to change for Go, but this
time the implementation is simpler than for JavaScript, since Go
supports multiple return values.