Go: Support arrays
Arrays.dfy now passes.  Arrays lean heavily on code in the runtime
library; this is only absolutely necessary because we need to remember
all the dimensions of an array, even if one of those dimensions is zero.
But it also makes generated code more concise and easier to debug by
cutting down on churn in converting between int and big.Int.

Somewhat unsatisfyingly, the array code relies heavily on the reflect
package, which is the only way to write code generically across
different array types.

As a hack to support modules, the IdProtect() method now shoves the
prefix  Go_ onto every identifier to make it public.  We can probably
be more precise than that.