Refactor generation of class members
Each language needs to keep track of different things during class
generation.
For instance, C# needs to put static members of traits in a special
place (the companion object),
and both Go and JavaScript handle fields differently from methods. The
base class was having to track all of these possible places, threading
all this state back through to CreateFunction(), CreateMethod(), and so
forth.  Now, instead we make CreateFunction() and CreateMethod() methods
of the new IClassWriter interface, so that each compiler class can be in
charge of what state needs to be remembered.