Go: Classes
The translation of the Dafny object system into Go is a little ...
subtle, but it seems to work so far.  Embedding a structure is *sort of*
like deriving from a class or trait, but *not really quite.* For
instance, we don't have upcasting, we have projection; but then we need
to be able to find the right "this" pointer after we project.  So the
solution is that a Dafny trait is a Go *struct* that *embeds* a Go
interface.  (The struct does make inheriting fields easy.)  The subclass
embeds the trait structure, then the constructor passes a pointer to the
new object as its own embedded
interface value.