Go: Iterators
Iterators.dfy now passes with Go.

The implementation method is fairly hacky: We run the iterator function
in a goroutine and use channels to signal when to resume and when the
iterator has yielded.  Goroutines are supposed to be lightweight, so
this should work decently.  The big issue is that garbage collection
won't cause the goroutine to stop or its channel to be closed (or GCed
itself).  I *believe* I've solved this using a finalizer, but finalizers
are themselves troublesome and I'm not sure how to test that my solution
actually helps.