Avoid generating an LHS in a separate statement in common cases
Many assignments were generating two statements along the lines of (in
Go):

    var _rhs int32 = foo
    lhs = _rhs

We now simply write

   lhs = foo

in these cases.  The one wrinkle is that Go now needs to explicitly
specify the type of a bitvector literal; previously, the declaration of
the type of rhs was enough to give the literal the right type.

An update to dafny0/RuntimeTypeTests0.dfy.expect was necessary because,
now that we generate fewer variables, there's one fewer error reported
during compilation.