Go: Coerce functions using anonymous functions
We were blithely assuming that we don't need to coerce between types
with the same head type, since almost all type parameters get erased,
but arrow types remain in their full glory.  This means we need to
coerce if, say, a function takes a type parameter A and a function from
int to A (since the function actually takes a function from int to
interface{}).