Go: Saner container APIs
We were having Seq.Index(), Array.Index(), etc. return refl.Values so
that they could be treated like lvalues.  Now they return pointers
instead.  This required generalizing the type of EmitArraySelect, since
before it could only affix something to the end of the array expression,
and we need to say

  *array.Index(...)

instead, putting the asterisk *before* the array.

While I was at it, I added public versions of Index() that take native
ints; there are a few places where that cuts down on the back-and-forth
between int and Int.