Go: Fold _module into main module; use program name as pkg if no Main()
The main package was just a silly little stub; now it contains the code
from the default module.  Also, not calling the package main is
important for a library, since if the package is called main, Go expects
an executable program with a main() method.  All this makes it more
reasonable to build a Go library from a Dafny library (though there
remains a dependency on the System_ package).