Go: Don't protect field names before capitalizing them
In generating the struct for a (co)datatype, we were using
Compiler.FormalName(), which uses IdProtect, and passing the result
through Capitalize.  When generating destructors, we were just
capitalizing, so we got the field name wrong.  Now we always bypass
IdProtect, which is unnecessary when we're going to capitalize anyway
(no capitalized identifiers conflict).

Added a test to comp/Dt.dfy to check the fix by giving fields names that
conflict with keywords in the different back-end languages.  Also added
a check that makes sure Go generates capitalized names for unnamed
fields.