Inline an LHS whenever it is stable
This defines a predicate IsStableExpr which is intended to be used in
forall compilation as well, where we want to generate a simple for loop
when the range is stable.  For now, we only use the function when
generating the LHS of an assignment; now LHSes such as a[i] will be
inlined, avoiding two temporary variables each time.