Compiler: Turn forall statement into simple loop when possible
There are many, many reasons why it may not be safe to run the
statements in a forall one by one.  But there are still enough common
safe cases that we can often generate a nice, simple loop to evaluate
them rather than generating data in one loop then doing the updates in
another.