Unix path separator in 'otherFilenames' not passed on
This commit fixes the following issue:

In cygwin and I assume other systems with Unix path separators, when Dafny is called with an external C# source file in a subdirectory, the C# compiler would not be passed the correct path. It would assume the external C# file is in the current directory.

e.g. `Dafny.exe src/file.dfy src/file-extern.dfy` would call the C# compiler with `file-extern.dfy` instead of `src/file-extern.dfy`.