Make `{:` a token instead of two separate tokens
This allows us to get rid of many `IF(IsAttribute())` manual
disambiguations.

Note that IsExistentialGuard needed to be updated to look for
lbracecolon instead of lbrace followed by colon.

Also note that strictly speaking, this is a breaking change:
Previously, `{ :myAttribute}` was accepted, but now it isn't any
more, and you have to remove the space.
But the whole test suite passes, so I think it's ok to enforce
a consistent style of not allowing a space between `{` and `:`.