New approach to disambiguate between attribute and set display:
Make `{:` a token instead of two separate tokens.
This looks like a neat thing to do, because it allows us to get
rid of many `IF(IsAttribute())` manual disambiguations.

Note that IsExistentialGuard needed to be updated to look for
lbracecolon instead of lbrace followed by colon.

Also note that strictly speaking, this is a breaking change:
Previously, `{ :myAttribute}` was accepted, but now it isn't any
more, and you have to remove the space.
But the whole test suite passes, so I think it's ok to enforce
a consistent style of not allowing a space between `{` and `:`.

However, we get more Coco warnings:
(Note that "deletable structure" seems to be Coco jargon for an optional
piece of syntax, i.e. what's written as "{ foo }" in Coco)

  LL1 warning in IteratorSpec: lbracecolon is start & successor of deletable structure
  LL1 warning in IteratorSpec: lbracecolon is start & successor of deletable structure
  LL1 warning in IteratorSpec: lbracecolon is start & successor of deletable structure
  LL1 warning in IteratorSpec: lbracecolon is start & successor of deletable structure
  LL1 warning in MethodSpec: lbracecolon is start & successor of deletable structure
  LL1 warning in MethodSpec: lbracecolon is start & successor of deletable structure
  LL1 warning in MethodSpec: lbracecolon is start & successor of deletable structure
  LL1 warning in MethodSpec: lbracecolon is start & successor of deletable structure
  LL1 warning in FunctionSpec: lbracecolon is start & successor of deletable structure
  LL1 warning in FunctionSpec: lbracecolon is start & successor of deletable structure
  LL1 warning in AssertStmt: lbracecolon is start & successor of deletable structure
  LL1 warning in AssumeStmt: lbracecolon is start & successor of deletable structure
  LL1 warning in ModifyStmt: lbracecolon is start & successor of deletable structure
  LL1 warning in LoopSpec: lbracecolon is start & successor of deletable structure
  LL1 warning in LoopSpec: lbracecolon is start & successor of deletable structure
  LL1 warning in LoopSpec: lbracecolon is start & successor of deletable structure

And these are to be taken seriously: They happen whenever a list of attributes
(which can be of length 0) precedes an Expression (which now can start with
attributes, too), so it's ambiguous whether the attributes belong to the thing
being parsed, or whether they belong to the Expression because the Expression
starts with a `{:attr} :|` or a `{:attr} :-`.