Use TargetWriter to format generated Java code
The only effect of this change is to change the formatting of the Java code that is generated. The `TargetWriter` class
has functionality for automatic indentation and is used elsewhere. So, I changed the code generation for
`Array<n>`, `ArrayInit<n>`, `Function<n>`, and `Tuple<n>` to write using `TargetWriter`.