Spend less time constructing BigInteger's in compiled code
In C#, create BigInteger's directly from int32, int64, or uint64 values, if possible (not just int32, as before),
and use the predefined BigInteger.Zero and BigInteger.One for 0 and 1.

In Go, use _dafny.One for 1 (analogously to what was already done for 0).

In JavaScript, define a ZERO and ONE in the run time and use these.
In JavaScript, also fixed two bugs:
* Fixed a three-times-occurring bug in DafnyRuntime.js where 2^53 was used as a `number` instead of as a `BigNumber`.
* Fixed numerical range of `number` among Dafny's pre-defined native types.

In Java, use ZERO or ONE or the constructor from a `long`, if possible. This led to a 25% speed-up in
one Dafny-implemented application.
Also, fixed bug in ULong-from-BigInteger constructor (and added comments) in class ULong in Dafny's run time.
Added comments in this file and added test cases that would have exposed the previous bug. Replaced MAXBI
constant by ULONG_LIMIT constant.