Fix ordering of static const's
Dafny does not care about the order in which static or module-level `const` declarations are given, but there must not be any cyclic
dependencies among them. The semantics is independent of the order of declaration. Target languages may care about this ordering,
so the compiler sorts the declarations according to the dependency relation computed in the Resolver. Alas, the comparison operation
used to sort the declarations was reversed, so the declarations ended up in the opposite order of what was needed.