Optimize comparisons to zero
As a C# example, x < BigInteger.Zero is slower than x.Sign < 0, possibly
because of the access of BigInteger.Zero (or perhaps just because
BigInteger.Sign is cheap).

Currently only implemented for C#, Java, and Go; the JavaScript BigInteger class we're using doesn't have a sign() method.  (We could use .isZero(), .isPositive(), and isNegative() in that case.)