Make DafnyRuntime.cs executable
It doesn't really make sense as an executable, but
every time it builds the mode is changed by msbuild, which is
an annoyance when working on Dafny.  There doesn't seem to
be an option to set the mode in the CopyFilesToOutputDirectory
section of a .csproj file.  It doesn't seem to me
that there's any harm in having it executable.