C#: Lazy sequence concatenation (#404)
* C#: Lazy sequence concatenation

Building a sequence of length n by repeatedly using + to append them takes O(n^2) time, and there isn't a convenient alternative to doing it this way (since Dafny doesn't have variable-length arrays).  This patch changes the C# runtime library to make concatenation lazy to knock this down to O(n) (so long as none of the intermediate sequences are ever accessed besides using +).

The behavior is very simple:  A sequence built by concatenation remains lazy until any operation besides + is performed on it, at which point all the sequences in the chain get concatenated in one go.  At that point, the sequence's elements are stored in an array as usual.

Caveats: There's all the usual laziness overhead---a little extra setup up front, and then an extra few checks on each access (including Select).  Also, Sequence is now an abstract class and all access to the underlying array goes through an abstract property, which may be slow.

* Hoist accesses to Elements out of loops

Elements being an abstract property, repeatedly accessing it was bogging down EqualsUntil (and hence Equals) considerably.

* C#: Track exact sequence length in ConcatSequence

* Fix typo

* C#: Exploit known length to avoid List in ConcatSequence.ComputeElements()