Treat receiver like other parameters in lit axioms
Previously, `this` was not treated like other parameters in the "lit axioms"
that allow function definitions to be expanded without consuming fuel in
the verifier. Perhaps that was okay before, we `this` was usually the same
in a recursive call. With type members for other types, like `datatype`,
a recursive function is likely to be called on something other than `this`.

In effect, this meant that the verifier could get stuck in infinite matching,
especially when faced with a false proof obligation (and in at least one
observed true proof obligation). The assertion in LitTriggers.dfy would
cause the verifier to never come back, but with this fix, the assertion is
quickly reported as an error.

This fix required `ThisExpr` nodes to be translated into Boogie AST nodes
with a filled-in type. Evidently, that was not needed before. For this reason,
this commit changed a number of calls to `ThisExpr` constructor.

Another parameter passed to Boogie functions is the heap. Since functions
do not change the heap, the heap parameter does not need to be lit-wrapped.
However, to reduce the chances that lit axioms will loop forever, dependencies
in the heap are disallowed--that is, if a function has a dependency on the
heap, it does not get lit axioms. This was done before. This could be improved
in two ways. (0) Checking an empty `reads` clause (which is what the code
does) does not detect dependencies on `const` fields. It is conceivable
that such dependencies could cause infinite matching of lit axioms, but I
have not seen any such example. (1) Instead of checking `reads` clauses,
perhaps the code could see if `decreases` expressions depend on the heap.
This would allow more examples to go through, and perhaps it could be done
without introducing infinite matching.

A final remark, the code previously had a special case: the nonemptiness of
the `reads` clause was ignored if the function signature contained an arrow
type. This let an example in TreeMapSimple.dfy go through. This special
case does not seem good, since the empty-`reads`-clause check can then
be subverted by introducing an unused arrow-typed parameter. So, I removed
this special case. I adjusted TreeMapSimple.dfy accordingly. Annoyingly, this
means the `TestY` assertion cannot be proved in `TestZ`--even with a manual
proof (because of lack of function extensionality, or as the case may be here,
lack of extensionality for functions with read effects).