Java: Allow extern classes to mix Dafny and extern code
If an extern class includes Dafny code, the generated code is now put into an abstract base class, whose name is prefixed with `_ExternBase_`, so that the user-provided class can extend it.

FIXME: Currently there is nothing to detect whether a class has *only* Dafny
methods.  Which means that any pure-Dafny module with an `{:extern}`
declaration (even just to give it an external name) requires the user to
provide a `__default` class (extending `_ExternBase___default`).

FIXME: The included tests break Go and JavaScript, which don't correctly
implement extern constructors.