Only consider a class an extern class if it has an extern member
It's useful to declare a module as {:extern} even if it has no extern members,
just to give it an externally-visible name.  Unfortunately, this causes its
default class to be considered (by Compiler.Compile()) to be extern as well,
and for Java this means forcing the user to provide an implementation of the
__default class that does nothing but extend _ExternBase___default.  This patch
makes sure that at least one member carries a zero- or one-argument {:extern}
before considering a class to be extern.

Why not a two-argument {:extern}?

    module {:extern "Library"} Library {
      static method Method() { ... }
      static method {:extern "Library.Class", "ExternMethod"} ExternMethod()
    }

Here the user has specified that the method Method lives not in
Library.__default but in Library.Class (presumably because __default looks
weird).  Since Library.__default has no methods that need to be filled in
by the user, it would be silly to force them to write

    package Library
    public class __default extends _ExternBase___default {
    }

Note that there is still a corner case:

    module {:extern "Library"} Library {
      class {:extern "Class"} Class {
        static method Method() { ... }
	static method {:extern "Library.Class.InnerClass", "ExternMethod"} ExternMethod()
      }
    }

Now we *do* need to generate _ExternBase_Class so that the user can supply
Library.Class so that they can define Library.Class.InnerClass.  It's not clear
this will ever come up, though (whereas the first example comes from the wild),
and in any case the workaround is just to define a dummy extern method.