Split out tests for extern constructors
Extern constructors are currently broken on Go and JavaScript, so I'm moving
that functionality from the tests in Extern.dfy and moving it to the new
ExternCtors.dfy, which only has RUN lines for C# and Java.  I've included the
extern code that *should* make the tests work (but currently doesn't).