Java: Unboxed primitive types
Primitive types are now represented by the actual primitives, in unboxed form.
Unsigned types are represented by their signed counterparts; UInt and friends
are no more (this was a previous patch).

To represent an array of unknown type, the dafny.Array<T> class is introduced.
It is *only* used when the type T is an unknown parameter, to avoid the
overhead of an object for each array.  (Multi-dimensional arrays always use the
corresponding dafny.Array{n} classes, since we need the object anyway to
remember the dimensions.)  The tricky thing here is coercing, say, byte[] to
dafny.Array<Byte> when calling a generic method expecting dafny.Array<T>, and
coercing in the other direction if a method returns a generic array.