Fixes; base DafnySequence on dafny.Array
Several fixes to the handling of arrays, especially their interaction with
sequences.  ByteArrayDafnySequence is now folded into ArrayDafnySequence, which
is now backed by a dafny.Array.  Also it now takes the element type descriptor
as an argument so that the correct array can be created (this will break extern
code).