Java: Dynamically find _type method of extern class
Since an extern class may be one whose source code the user can't change, and
since the default value of a class is almost always null so an explicit _type
function isn't necessary, Java now does what the other backends do and
dynamically looks for a static _type method for any extern class.  This may be
a performance issue.  TODO: Consider a WeakHashMap from java.lang.Class to
dafny.Type (or some sort of dafny.TypeFunction in the case of a generic type)
to save on reflection.