DafnySequence performance tweaks
A couple of small issues look to be having outsized influence: Create() loops
over BigInteger and converts each to an int rather than looping over ints and
creating a BigInteger for each; the latter is much faster.  Also,
LazyDafnySequence doesn't override verbatimString(), which was causing
concatenated DafnySequence<Character>s to be very slow in converting to
strings.