feat: Body-less loops (#513)
Compute loop targets as those free mutable variables (including the
heap) that occur in the guard, invariant, or decreases clause. If the
loop has a modifies clause--whatever the modifies clause says--then
the heap is counted as a free loop target.

Verify context of body-less loop based on loop specification and
the computed loop targets.

Change loop-target hovertext to be part of the "no body" warning.

fix: body-less loops are not always ghost.

fix: `unchanged` expressions use the current heap.