Introduce Type objects to Java runtime
The system for using strings as runtime descriptors was buggy, and this paves
the way for using RTDs to do things like intelligently (and quickly) allocate
arrays that may be primitive when we get rid of the use of boxed primitives.

Also fixed a Go bug (exposed by a new test): When a zero was emitted as a
default value, it wasn't qualified with its type, which broke the generation of
runtime type descriptors for primitives (since the Default() method must return
an interface{}, which meant the 0 defaulted to int every time).