feat: allow more PropagateFailure... members
Previously, the `:-` construct would always use the method or function `PropagateFailure()`
when propagating a failure. Now, a type can provide several propagate-failure members with
different result types. This makes it possible in some cases to convert between types in
failure cases.

In more detail, the extended functionality applies only to `:-` statements, not `:-` expressions.
Consider a `:-` statement whose right-hand side has type `G<X>`, where the enclosing method has
one out-parameter, say of type `H<Y>`.

This will now consider all functions and one-out-parameter methods in type `G<X>` whose name
starts with `PropagateFailure`. Consider such a propagate-failure member to be "suited" if its return
type, say `J<Z>`, has the same base-type declaration as `H<Y>`. That is, keep expanding `J<Z>`
along type synonyms and subset-type base types, yielding a type `JJ<ZZ>`, and do the same for
`H<Y>`, yielding `HH<YY>`. Then, the propagate-failure member is "suitable" if `JJ` is the same as `HH`.

If there's a unique suitable propagate-failure member, then it is picked. If there's more than one suitable
propagate-failure member, then an error is generated. If there's no suitable propagate-failure member,
then the plainly named `PropagateFailure()` is picked.

If there is no plainly named `PropagateFailure()` member, then an error is generated, even if there is
a suitable propagate-failure member.

Design note: Since Dafny does not have overloading of user-declared members, this name-prefix trick
appears useful. However, there are no other such name-prefix tricks in Dafny, so the design of this
extension is dubious. Also, the conversion has to be declared in the source type--that is, the type of
the right-hand side of the `:-`. This may not always be the best design choice. Nevertheless, in order
to experiment with ways to extend the current `:-` behavior, this change seems like a good step. It is
backward compatible, and I hope it leads to experimentation that will better inform what a more desirable
extension would be.