Use java.lang.Object instead of dafny.Array for generic arrays
There was a fatal flaw in the scheme that encoded an array of generic type T as
a `dafny.Array<T>` and an array of known type as a plain array: While we can
unwrap a returned array as necessary to coerce from, say,
`dafny.Array<Integer>` to `int[]`, we cannot expect to do the same for a
`dafny.Array<T>` somewhere inside a class or datatype.  The only alternative
that doesn't add overhead to the known-element-type case is to represent an
array of a type parameter as a `java.lang.Object` (since, sadly, primitive
arrays and object arrays share no other superclass).  Generated code will still
be type-correct since the compiler knows which `java.lang.Object`s are arrays
of which type, and we can use the `dafny.Type<T>` object passed into a generic
method to perform array operations.

The `dafny.Array` class is retained for convenience in programming but is no
longer used by generated code.