Introduce lvalue abstraction (#523)
* Introduce lvalue abstraction

An lvalue being something that can be on the left of the assignment operator,
an ILvalue is something we know how to render either as a value being read or a
location being written to.  Having this abstraction means that a backend can
specify that a particular expression must be assigned to using something
besides an actual assignment statement; this way a call to EmitAssignment may
not literally emit an assignment.  For example, the Java backend was not
properly implementing setter methods (see the included test case), but now
they're easy to implement using a custom ILvalue implementation (see
JavaCompiler.GetterSetterLvalue) without changes to the base class.  (The
MemberSelectObjIsTrait parameter to EmitMemberSelect, which was only there for
the Java backend, is no more.)

In particular, this paves the way for Java to use unboxed primitives, since
representing an array of T as T[] forces us to use arrays of boxed types, but
using a plain Object means we can't index into the array as an lvalue---we have
to call a method (such as one in java.lang.reflect.Array, though we'll use
something faster) instead of an assignment statement.

* Fix regression introduced by ILvalue refactor