Java: Unboxed primitives (#472)
* Java: Unboxed primitive types

Primitive types are now represented by the actual primitives, in unboxed form.
Unsigned types are represented by their signed counterparts; UInt and friends
are no more (this was a previous patch).

To represent an array of unknown type, the dafny.Array<T> class is introduced.
It is *only* used when the type T is an unknown parameter, to avoid the
overhead of an object for each array.  (Multi-dimensional arrays always use the
corresponding dafny.Array{n} classes, since we need the object anyway to
remember the dimensions.)  The tricky thing here is coercing, say, byte[] to
dafny.Array<Byte> when calling a generic method expecting dafny.Array<T>, and
coercing in the other direction if a method returns a generic array.

* Fixes; base DafnySequence on dafny.Array

Several fixes to the handling of arrays, especially their interaction with
sequences.  ByteArrayDafnySequence is now folded into ArrayDafnySequence, which
is now backed by a dafny.Array.  Also it now takes the element type descriptor
as an argument so that the correct array can be created (this will break extern
code).

* Java, Go: Fix conversion from character to native int

* Java: Dynamically find _type method of extern class

Since an extern class may be one whose source code the user can't change, and
since the default value of a class is almost always null so an explicit _type
function isn't necessary, Java now does what the other backends do and
dynamically looks for a static _type method for any extern class.  This may be
a performance issue.  TODO: Consider a WeakHashMap from java.lang.Class to
dafny.Type (or some sort of dafny.TypeFunction in the case of a generic type)
to save on reflection.

* DafnySequence performance tweaks

A couple of small issues look to be having outsized influence: Create() loops
over BigInteger and converts each to an int rather than looping over ints and
creating a BigInteger for each; the latter is much faster.  Also,
LazyDafnySequence doesn't override verbatimString(), which was causing
concatenated DafnySequence<Character>s to be very slow in converting to
strings.

* Avoid assertion error on ghost method with {:extern} declaration

* Java: Update test in runtime library

* Use java.lang.Object instead of dafny.Array for generic arrays

There was a fatal flaw in the scheme that encoded an array of generic type T as
a `dafny.Array<T>` and an array of known type as a plain array: While we can
unwrap a returned array as necessary to coerce from, say,
`dafny.Array<Integer>` to `int[]`, we cannot expect to do the same for a
`dafny.Array<T>` somewhere inside a class or datatype.  The only alternative
that doesn't add overhead to the known-element-type case is to represent an
array of a type parameter as a `java.lang.Object` (since, sadly, primitive
arrays and object arrays share no other superclass).  Generated code will still
be type-correct since the compiler knows which `java.lang.Object`s are arrays
of which type, and we can use the `dafny.Type<T>` object passed into a generic
method to perform array operations.

The `dafny.Array` class is retained for convenience in programming but is no
longer used by generated code.

* Fixes